SerialComAnalyzer/requirements.txt
pyqt6==6.1.1
pyserial==3.5